requirements.txt
amqp==5.0.5
arrow==1.0.3
asgiref==3.3.1
billiard==3.6.3.0
blessed==1.18.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.6
django-allauth==0.44.0
django-celery-beat==2.2.0
django-crispy-forms==1.11.1
django-crontab==0.7.1
django-localflavor==3.0.1
django-picklefield==3.0.1
django-q==1.3.5
django-timezone-field==4.1.1
kombu==5.0.2
oauthlib==3.1.0
Pillow==8.1.0
PyJWT==2.0.1
python-crontab==2.5.1
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.56.0
vine==5.0.0
